qwen35: fix GatedDeltaNet conv direction + gated-norm gate; fp16 kernel storage
- diff against transformers Qwen3Next as oracle revealed two shared biases:
  * depthwise causal conv1d taps were reversed relative to F.conv1d
    (padding=kernel-1) orientation
  * linear-attn output norm is weight*x_norm*silu(z), not (1+weight)
- fix cpu.rs / qwen35.cuh (prefill + decode conv_step) / gpu.rs norm upload
- CUDA elementwise/prefill/decode kernels now store activations in fp16
  (__half) instead of bf16 for cleaner GEMM I/O
- long-context validation examples dump multi-position logits
- verified: 8-token probe argmax 310 (== HF-formula oracle), p10 greedy
  KEY-P10-767211, all 6 public functional cases pass 6/6

+++ crates/apxinf-cuda/adapters/custom_kernels.cu
@@ -27,6 +27,7 @@ extern "C" cudaError_t apxinf_static_evict_l2(
   if (buffer == nullptr || bytes < sizeof(uint32_t) ||
       bytes % sizeof(uint32_t) != 0) {
     return cudaErrorInvalidValue;
+
   }
   constexpr int threads = 256;
   int blocks = static_cast<int>((bytes / sizeof(uint32_t) + threads - 1) /
@@ -423,9 +424,9 @@ extern "C" cudaError_t apxinf_qwen_dequant_w4a16_bf16(
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_dequant_w4a16_bf16_kernel<<<blocks, threads, 0, stream>>>(
       static_cast<const int32_t*>(packed),
-      static_cast<const __nv_bfloat16*>(scale),
+      static_cast<const __half*>(scale),
       static_cast<const int32_t*>(zp),
-      static_cast<__nv_bfloat16*>(out), out_dim, in_dim);
+      static_cast<__half*>(out), out_dim, in_dim);
   return cudaGetLastError();
 }
 
@@ -436,9 +437,9 @@ extern "C" cudaError_t apxinf_qwen_mul_bf16(
   int64_t blocks64 = (n + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_mul_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(a),
-      static_cast<const __nv_bfloat16*>(b),
-      static_cast<__nv_bfloat16*>(out), n);
+      static_cast<const __half*>(a),
+      static_cast<const __half*>(b),
+      static_cast<__half*>(out), n);
   return cudaGetLastError();
 }
 
@@ -449,8 +450,8 @@ extern "C" cudaError_t apxinf_qwen_silu_bf16(
   int64_t blocks64 = (n + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_silu_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(a),
-      static_cast<__nv_bfloat16*>(out), n);
+      static_cast<const __half*>(a),
+      static_cast<__half*>(out), n);
   return cudaGetLastError();
 }
 
@@ -461,9 +462,9 @@ extern "C" cudaError_t apxinf_qwen_sigmoid_mul_bf16(
   int64_t blocks64 = (n + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_sigmoid_mul_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(a),
-      static_cast<const __nv_bfloat16*>(b),
-      static_cast<__nv_bfloat16*>(out), n);
+      static_cast<const __half*>(a),
+      static_cast<const __half*>(b),
+      static_cast<__half*>(out), n);
   return cudaGetLastError();
 }
 
@@ -474,8 +475,8 @@ extern "C" cudaError_t apxinf_qwen_accum_bf16(
   int64_t blocks64 = (n + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_accum_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<__nv_bfloat16*>(dst),
-      static_cast<const __nv_bfloat16*>(src), n);
+      static_cast<__half*>(dst),
+      static_cast<const __half*>(src), n);
   return cudaGetLastError();
 }
 
@@ -486,9 +487,9 @@ extern "C" cudaError_t apxinf_qwen_rms_norm_bf16(
   int threads = 256;
   size_t shmem = (size_t)threads * sizeof(float);
   qwen_rms_norm_bf16_kernel<<<rows, threads, shmem, stream>>>(
-      static_cast<const __nv_bfloat16*>(x),
-      static_cast<const __nv_bfloat16*>(w),
-      static_cast<__nv_bfloat16*>(out), rows, cols, eps);
+      static_cast<const __half*>(x),
+      static_cast<const __half*>(w),
+      static_cast<__half*>(out), rows, cols, eps);
   return cudaGetLastError();
 }
 
@@ -500,9 +501,9 @@ extern "C" cudaError_t apxinf_qwen_qg_split_bf16(
   int64_t blocks64 = (total + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_qg_split_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(qg),
-      static_cast<__nv_bfloat16*>(q),
-      static_cast<__nv_bfloat16*>(gate), total, heads, hd);
+      static_cast<const __half*>(qg),
+      static_cast<__half*>(q),
+      static_cast<__half*>(gate), total, heads, hd);
   return cudaGetLastError();
 }
 
@@ -514,9 +515,9 @@ extern "C" cudaError_t apxinf_qwen_partial_rope_bf16(
   int64_t blocks64 = (pairs + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_partial_rope_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<__nv_bfloat16*>(x),
-      static_cast<const __nv_bfloat16*>(cos),
-      static_cast<const __nv_bfloat16*>(sin),
+      static_cast<__half*>(x),
+      static_cast<const __half*>(cos),
+      static_cast<const __half*>(sin),
       pairs, heads, hd, half, static_cast<const int*>(pos0));
   return cudaGetLastError();
 }
@@ -530,9 +531,9 @@ extern "C" cudaError_t apxinf_qwen_conv_silu_bf16(
   int64_t blocks64 = (total + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_conv_silu_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(x),
-      static_cast<const __nv_bfloat16*>(w),
-      static_cast<__nv_bfloat16*>(out), L, conv_dim);
+      static_cast<const __half*>(x),
+      static_cast<const __half*>(w),
+      static_cast<__half*>(out), L, conv_dim);
   return cudaGetLastError();
 }
 
@@ -545,13 +546,13 @@ extern "C" cudaError_t apxinf_qwen_delta_recurrence_bf16(
   if (vd > 128 || kd % DR_KDCHUNK != 0) return cudaErrorInvalidValue;
   dim3 block(vd, DR_KDCHUNK);
   qwen_delta_recurrence_bf16_kernel<<<nv, block, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(q),
-      static_cast<const __nv_bfloat16*>(k),
-      static_cast<const __nv_bfloat16*>(v),
-      static_cast<const __nv_bfloat16*>(beta),
-      static_cast<const __nv_bfloat16*>(g),
+      static_cast<const __half*>(q),
+      static_cast<const __half*>(k),
+      static_cast<const __half*>(v),
+      static_cast<const __half*>(beta),
+      static_cast<const __half*>(g),
       static_cast<float*>(state),
-      static_cast<__nv_bfloat16*>(out),
+      static_cast<__half*>(out),
       L, nv, kd, vd);
   return cudaGetLastError();
 }
@@ -563,11 +564,11 @@ extern "C" cudaError_t apxinf_qwen_attention_bf16(
   if (L <= 0 || heads <= 0 || kvheads <= 0 || hd <= 0) return cudaErrorInvalidValue;
   if (hd > 256 || (hd & (hd - 1)) != 0) return cudaErrorInvalidValue;
   qwen_attention_bf16_kernel<<<L * heads, hd, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(q),
-      static_cast<const __nv_bfloat16*>(k),
-      static_cast<const __nv_bfloat16*>(v),
-      static_cast<const __nv_bfloat16*>(gate),
-      static_cast<__nv_bfloat16*>(out),
+      static_cast<const __half*>(q),
+      static_cast<const __half*>(k),
+      static_cast<const __half*>(v),
+      static_cast<const __half*>(gate),
+      static_cast<__half*>(out),
       L, heads, kvheads, hd);
   return cudaGetLastError();
 }
@@ -582,10 +583,10 @@ extern "C" cudaError_t apxinf_qwen_conv_split_bf16(
   int64_t b64 = (total + threads - 1) / threads;
   int blocks = b64 > (1 << 20) ? (1 << 20) : (int)b64;
   qwen_conv_split_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(conv),
-      static_cast<__nv_bfloat16*>(q),
-      static_cast<__nv_bfloat16*>(k),
-      static_cast<__nv_bfloat16*>(v),
+      static_cast<const __half*>(conv),
+      static_cast<__half*>(q),
+      static_cast<__half*>(k),
+      static_cast<__half*>(v),
       L, nk, nv, kd, vd, conv_dim);
   return cudaGetLastError();
 }
@@ -597,8 +598,8 @@ extern "C" cudaError_t apxinf_qwen_l2norm_bf16(
   int threads = 256;
   size_t shmem = (size_t)threads * sizeof(float);
   qwen_l2norm_bf16_kernel<<<rows, threads, shmem, stream>>>(
-      static_cast<const __nv_bfloat16*>(x),
-      static_cast<__nv_bfloat16*>(out), rows, cols, eps, scale);
+      static_cast<const __half*>(x),
+      static_cast<__half*>(out), rows, cols, eps, scale);
   return cudaGetLastError();
 }
 
@@ -608,11 +609,11 @@ extern "C" cudaError_t apxinf_qwen_gemm_w4a16_bf16(
   if (m <= 0 || n <= 0 || k <= 0) return cudaErrorInvalidValue;
   dim3 grid((n + QW4_BN - 1) / QW4_BN, (m + QW4_BM - 1) / QW4_BM);
   qwen_gemm_w4a16_bf16_kernel<<<grid, 256, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(a),
+      static_cast<const __half*>(a),
       static_cast<const int32_t*>(packed),
-      static_cast<const __nv_bfloat16*>(scale),
+      static_cast<const __half*>(scale),
       static_cast<const int32_t*>(zp),
-      static_cast<__nv_bfloat16*>(c), m, n, k);
+      static_cast<__half*>(c), m, n, k);
   return cudaGetLastError();
 }
 
@@ -624,12 +625,12 @@ extern "C" cudaError_t apxinf_qwen_beta_g_bf16(
   int64_t blocks64 = ((int64_t)total + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_beta_g_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(a),
-      static_cast<const __nv_bfloat16*>(b),
+      static_cast<const __half*>(a),
+      static_cast<const __half*>(b),
       static_cast<const float*>(a_log),
       static_cast<const float*>(dt_bias),
-      static_cast<__nv_bfloat16*>(beta),
-      static_cast<__nv_bfloat16*>(g), total, nv);
+      static_cast<__half*>(beta),
+      static_cast<__half*>(g), total, nv);
   return cudaGetLastError();
 }
 
@@ -640,8 +641,8 @@ extern "C" cudaError_t apxinf_qwen_copy_bf16(
   int64_t blocks64 = (n + threads - 1) / threads;
   int blocks = blocks64 > (1 << 20) ? (1 << 20) : (int)blocks64;
   qwen_copy_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(src),
-      static_cast<__nv_bfloat16*>(dst), n);
+      static_cast<const __half*>(src),
+      static_cast<__half*>(dst), n);
   return cudaGetLastError();
 }
 
@@ -652,11 +653,11 @@ extern "C" cudaError_t apxinf_qwen_attention_decode_bf16(
   if (heads <= 0 || kvheads <= 0 || hd <= 0) return cudaErrorInvalidValue;
   if (hd > 256 || (hd & (hd - 1)) != 0) return cudaErrorInvalidValue;
   qwen_attention_decode_bf16_kernel<<<heads, hd, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(q),
-      static_cast<const __nv_bfloat16*>(kcache),
-      static_cast<const __nv_bfloat16*>(vcache),
-      static_cast<const __nv_bfloat16*>(gate),
-      static_cast<__nv_bfloat16*>(out),
+      static_cast<const __half*>(q),
+      static_cast<const __half*>(kcache),
+      static_cast<const __half*>(vcache),
+      static_cast<const __half*>(gate),
+      static_cast<__half*>(out),
       static_cast<const int*>(seq_ptr), heads, kvheads, hd);
   return cudaGetLastError();
 }
@@ -668,10 +669,10 @@ extern "C" cudaError_t apxinf_qwen_conv_step_silu_bf16(
   int threads = 256;
   int blocks = (conv_dim + threads - 1) / threads;
   qwen_conv_step_silu_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(cur),
-      static_cast<__nv_bfloat16*>(hist),
-      static_cast<const __nv_bfloat16*>(w),
-      static_cast<__nv_bfloat16*>(out), conv_dim);
+      static_cast<const __half*>(cur),
+      static_cast<__half*>(hist),
+      static_cast<const __half*>(w),
+      static_cast<__half*>(out), conv_dim);
   return cudaGetLastError();
 }
 
@@ -683,13 +684,13 @@ extern "C" cudaError_t apxinf_qwen_delta_step_bf16(
   if (vd > 128 || kd % DR_KDCHUNK != 0) return cudaErrorInvalidValue;
   dim3 block(vd, DR_KDCHUNK);
   qwen_delta_step_bf16_kernel<<<nv, block, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(q),
-      static_cast<const __nv_bfloat16*>(k),
-      static_cast<const __nv_bfloat16*>(v),
-      static_cast<const __nv_bfloat16*>(beta),
-      static_cast<const __nv_bfloat16*>(g),
+      static_cast<const __half*>(q),
+      static_cast<const __half*>(k),
+      static_cast<const __half*>(v),
+      static_cast<const __half*>(beta),
+      static_cast<const __half*>(g),
       static_cast<float*>(state),
-      static_cast<__nv_bfloat16*>(out), nv, kd, vd);
+      static_cast<__half*>(out), nv, kd, vd);
   return cudaGetLastError();
 }
 
@@ -700,8 +701,8 @@ extern "C" cudaError_t apxinf_qwen_copy_at_bf16(
   int threads = 256;
   int blocks = (n + threads - 1) / threads;
   qwen_copy_at_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(src),
-      static_cast<__nv_bfloat16*>(dst_base),
+      static_cast<const __half*>(src),
+      static_cast<__half*>(dst_base),
       static_cast<const int*>(pos_ptr), stride_elems, n);
   return cudaGetLastError();
 }
@@ -713,10 +714,10 @@ extern "C" cudaError_t apxinf_qwen_gemm_w4a16_m1_bf16(
   int threads = 128;
   int blocks = (n + threads - 1) / threads;
   qwen_gemm_w4a16_m1_bf16_kernel<<<blocks, threads, 0, stream>>>(
-      static_cast<const __nv_bfloat16*>(a),
+      static_cast<const __half*>(a),
       static_cast<const int32_t*>(packed),
-      static_cast<const __nv_bfloat16*>(scale),
+      static_cast<const __half*>(scale),
       static_cast<const int32_t*>(zp),
-      static_cast<__nv_bfloat16*>(c), n, k);
+      static_cast<__half*>(c), n, k);
   return cudaGetLastError();
 }
